migrate: backend, dummy_core, targets
