+++ crates/moon/tests/test_cases/backend/flag/lib/hello.mbt
