+++ crates/moon/tests/test_cases/backend/flag/lib/hello_wbtest.mbt
