+++ crates/moon/tests/test_cases/backend/flag/lib/moon.pkg.json
