+++ crates/moon/tests/test_cases/backend/js_format/lib0/moon.pkg.json
