+++ crates/moon/tests/test_cases/backend/js_format/lib1/hello_wbtest.mbt
