+++ crates/moon/tests/test_cases/backend/js_format/lib2/hello_wbtest.mbt
