+++ crates/moon/tests/test_cases/backend/js_format/lib3/hello_wbtest.mbt
