+++ crates/moon/tests/test_cases/backend/js_format/lib3/moon.pkg.json
