+++ crates/moon/tests/test_cases/dummy_core/0/moon.pkg.json
