+++ crates/moon/tests/test_cases/dummy_core/1/moon.pkg.json
