+++ crates/moon/tests/test_cases/dummy_core/2/moon.pkg.json
