+++ crates/moon/tests/test_cases/dummy_core/char/moon.pkg.json
