+++ crates/moon/tests/test_cases/dummy_core/iter/moon.pkg.json
