+++ crates/moon/tests/test_cases/dummy_core/moon.mod.json
