+++ crates/moon/tests/test_cases/dummy_core/prelude/moon.pkg.json
