+++ crates/moon/tests/test_cases/hello/mod.rs
@@ -18,3 +18,80 @@ fn test_hello() {
         expect!["false"],
     );
 }
+
+#[test]
+fn test_source_map() {
+    let dir = TestDir::new("hello");
+
+    // no -source-map in wasm backend
+    check(
+        get_stdout(
+            &dir,
+            [
+                "build",
+                "--target",
+                "wasm",
+                "--debug",
+                "--dry-run",
+                "--sort-input",
+            ],
+        ),
+        expect![[r#"
+            moonc build-package ./main/main.mbt -o ./target/wasm/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm/release/bundle -pkg-sources hello/main:./main -target wasm -g -O0
+            moonc link-core $MOON_HOME/lib/core/target/wasm/release/bundle/core.core ./target/wasm/debug/build/main/main.core -main hello/main -o ./target/wasm/debug/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm -g -O0
+        "#]],
+    );
+    check(
+        get_stdout(
+            &dir,
+            [
+                "build",
+                "--target",
+                "wasm-gc",
+                "--debug",
+                "--dry-run",
+                "--sort-input",
+            ],
+        ),
+        expect![[r#"
+            moonc build-package ./main/main.mbt -o ./target/wasm-gc/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources hello/main:./main -target wasm-gc -g -O0 -source-map
+            moonc link-core $MOON_HOME/lib/core/target/wasm-gc/release/bundle/core.core ./target/wasm-gc/debug/build/main/main.core -main hello/main -o ./target/wasm-gc/debug/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm-gc -g -O0 -source-map
+        "#]],
+    );
+    check(
+        get_stdout(
+            &dir,
+            [
+                "build",
+                "--target",
+                "js",
+                "--debug",
+                "--dry-run",
+                "--sort-input",
+            ],
+        ),
+        expect![[r#"
+            moonc build-package ./main/main.mbt -o ./target/js/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources hello/main:./main -target js -g -O0 -source-map
+            moonc link-core $MOON_HOME/lib/core/target/js/release/bundle/core.core ./target/js/debug/build/main/main.core -main hello/main -o ./target/js/debug/build/main/main.js -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target js -g -O0 -source-map
+        "#]],
+    );
+}
+
+#[test]
+fn test_find_ancestor_with_mod() {
+    let dir = TestDir::new("hello");
+
+    check(
+        get_stdout(&dir, ["run", "main"]),
+        expect![[r#"
+            Hello, world!
+        "#]],
+    );
+
+    check(
+        get_stdout(&dir, ["run", "main"]),
+        expect![[r#"
+            Hello, world!
+        "#]],
+    );
+}
