+++ crates/moon/tests/test_cases/mod.rs
@@ -29,10 +29,12 @@ use moonutil::{
 use serde::{Deserialize, Serialize};
 use walkdir::WalkDir;
 
+mod backend;
 mod backend_config;
 mod design;
 mod diamond_pkg;
 mod docs_examples;
+mod dummy_core;
 mod expect_test;
 mod extra_flags;
 mod fancy_import;
@@ -49,6 +51,7 @@ mod output_format;
 mod packages;
 mod simple_pkg;
 mod target_backend;
+mod targets;
 mod test_error_report;
 mod test_filter;
 
@@ -62,1304 +65,6 @@ fn test_need_link() {
             moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg username/hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources username/hello/main:./main -target wasm-gc
             moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main username/hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/main:./main -target wasm-gc
             moonc link-core ./target/wasm-gc/release/build/lib/lib.core -main username/hello/lib -o ./target/wasm-gc/release/build/lib/lib.wasm -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -target wasm-gc
-        "#]],
-    );
-}
-#[test]
-fn test_dummy_core() {
-    let test_dir = TestDir::new("dummy-core.in");
-    let dir = dunce::canonicalize(test_dir.as_ref()).unwrap();
-
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["check", "--sort-input"])
-        .assert()
-        .success();
-
-    #[cfg(unix)]
-    {
-        let p = dir.join("target/packages.json");
-        check(
-            replace_dir(&std::fs::read_to_string(p).unwrap(), &dir),
-            expect![[r#"
-                {
-                  "source_dir": "$ROOT",
-                  "name": "moonbitlang/core",
-                  "packages": [
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/0",
-                      "root": "moonbitlang/core",
-                      "rel": "0",
-                      "files": {
-                        "$ROOT/0/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/0/y.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {
-                        "$ROOT/0/y_wbtest.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/0/0.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/1",
-                      "root": "moonbitlang/core",
-                      "rel": "1",
-                      "files": {
-                        "$ROOT/1/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/1/x.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/1/x.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/1/x.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {
-                        "$ROOT/1/x_wbtest.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/1/1.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/2",
-                      "root": "moonbitlang/core",
-                      "rel": "2",
-                      "files": {
-                        "$ROOT/2/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/1",
-                          "alias": "1",
-                          "fspath": "$ROOT/1"
-                        }
-                      ],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/2/2.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/char",
-                      "root": "moonbitlang/core",
-                      "rel": "char",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/coverage",
-                          "alias": "coverage",
-                          "fspath": "$ROOT/coverage"
-                        }
-                      ],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/char/char.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/coverage",
-                      "root": "moonbitlang/core",
-                      "rel": "coverage",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/coverage/coverage.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/iter",
-                      "root": "moonbitlang/core",
-                      "rel": "iter",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/coverage",
-                          "alias": "coverage",
-                          "fspath": "$ROOT/coverage"
-                        }
-                      ],
-                      "wbtest-deps": [
-                        {
-                          "path": "moonbitlang/core/char",
-                          "alias": "char",
-                          "fspath": "$ROOT/char"
-                        }
-                      ],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/iter/iter.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/prelude",
-                      "root": "moonbitlang/core",
-                      "rel": "prelude",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/wasm-gc/release/check/prelude/prelude.mi"
-                    }
-                  ],
-                  "deps": [],
-                  "backend": "wasm-gc",
-                  "opt_level": "release",
-                  "source": null
-                }"#]],
-        );
-    }
-    snapbox::cmd::Command::new(moon_bin())
-        .current_dir(&dir)
-        .args(["check", "--target", "js", "--sort-input"])
-        .assert()
-        .success();
-
-    #[cfg(unix)]
-    {
-        let p = dir.join("target/packages.json");
-        check(
-            replace_dir(&std::fs::read_to_string(p).unwrap(), &dir),
-            expect![[r#"
-                {
-                  "source_dir": "$ROOT",
-                  "name": "moonbitlang/core",
-                  "packages": [
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/0",
-                      "root": "moonbitlang/core",
-                      "rel": "0",
-                      "files": {
-                        "$ROOT/0/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/0/y.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {
-                        "$ROOT/0/y_wbtest.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/0/y_wbtest.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/0/0.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/1",
-                      "root": "moonbitlang/core",
-                      "rel": "1",
-                      "files": {
-                        "$ROOT/1/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        },
-                        "$ROOT/1/x.js.mbt": {
-                          "backend": [
-                            "Js"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/1/x.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        },
-                        "$ROOT/1/x.wasm.mbt": {
-                          "backend": [
-                            "Wasm"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {
-                        "$ROOT/1/x_wbtest.wasm-gc.mbt": {
-                          "backend": [
-                            "WasmGC"
-                          ],
-                          "optlevel": [
-                            "Release",
-                            "Debug"
-                          ]
-                        }
-                      },
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/1/1.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/2",
-                      "root": "moonbitlang/core",
-                      "rel": "2",
-                      "files": {
-                        "$ROOT/2/lib.mbt": {
-                          "backend": [
-                            "Wasm",
-                            "WasmGC",
-                            "Js",
-                            "Native",
-                            "LLVM"
-                          ],
-                          "optlevel": [
-                            "Debug",
-                            "Release"
-                          ]
-                        }
-                      },
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/1",
-                          "alias": "1",
-                          "fspath": "$ROOT/1"
-                        }
-                      ],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/2/2.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/char",
-                      "root": "moonbitlang/core",
-                      "rel": "char",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/coverage",
-                          "alias": "coverage",
-                          "fspath": "$ROOT/coverage"
-                        }
-                      ],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/char/char.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/coverage",
-                      "root": "moonbitlang/core",
-                      "rel": "coverage",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/coverage/coverage.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/iter",
-                      "root": "moonbitlang/core",
-                      "rel": "iter",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [
-                        {
-                          "path": "moonbitlang/core/coverage",
-                          "alias": "coverage",
-                          "fspath": "$ROOT/coverage"
-                        }
-                      ],
-                      "wbtest-deps": [
-                        {
-                          "path": "moonbitlang/core/char",
-                          "alias": "char",
-                          "fspath": "$ROOT/char"
-                        }
-                      ],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/iter/iter.mi"
-                    },
-                    {
-                      "is-main": false,
-                      "is-third-party": false,
-                      "root-path": "$ROOT/prelude",
-                      "root": "moonbitlang/core",
-                      "rel": "prelude",
-                      "files": {},
-                      "wbtest-files": {},
-                      "test-files": {},
-                      "mbt-md-files": {},
-                      "deps": [],
-                      "wbtest-deps": [],
-                      "test-deps": [
-                        {
-                          "path": "moonbitlang/core/prelude",
-                          "alias": "prelude",
-                          "fspath": "$ROOT/prelude"
-                        }
-                      ],
-                      "artifact": "$ROOT/target/js/release/check/prelude/prelude.mi"
-                    }
-                  ],
-                  "deps": [],
-                  "backend": "js",
-                  "opt_level": "release",
-                  "source": null
-                }"#]],
-        );
-    };
-
-    check(
-        get_stdout(&dir, ["check", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc check -o ./target/wasm-gc/release/check/prelude/prelude.mi -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/coverage/coverage.mi -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/iter/iter.mi -pkg moonbitlang/core/iter -i ./target/wasm-gc/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/char/char.mi -pkg moonbitlang/core/char -i ./target/wasm-gc/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc
-            moonc check ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/check/1/1.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc check ./2/lib.mbt -o ./target/wasm-gc/release/check/2/2.mi -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/check/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc check ./1/lib.mbt ./1/x.wasm-gc.mbt ./1/x_wbtest.wasm-gc.mbt -o ./target/wasm-gc/release/check/1/1.whitebox_test.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm-gc.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm-gc.mbt -o ./target/wasm-gc/release/check/0/0.whitebox_test.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/check/0/0.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["check", "--dry-run", "--target", "wasm", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc check -o ./target/wasm/release/check/prelude/prelude.mi -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm
-            moonc check -o ./target/wasm/release/check/coverage/coverage.mi -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm
-            moonc check -o ./target/wasm/release/check/iter/iter.mi -pkg moonbitlang/core/iter -i ./target/wasm/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm
-            moonc check -o ./target/wasm/release/check/char/char.mi -pkg moonbitlang/core/char -i ./target/wasm/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm
-            moonc check ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/release/check/1/1.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm
-            moonc check ./2/lib.mbt -o ./target/wasm/release/check/2/2.mi -pkg moonbitlang/core/2 -i ./target/wasm/release/check/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm
-            moonc check ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/release/check/1/1.whitebox_test.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm.mbt -o ./target/wasm/release/check/0/0.whitebox_test.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm.mbt -o ./target/wasm/release/check/0/0.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["check", "--dry-run", "--target", "wasm-gc", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc check -o ./target/wasm-gc/release/check/prelude/prelude.mi -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/coverage/coverage.mi -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/iter/iter.mi -pkg moonbitlang/core/iter -i ./target/wasm-gc/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc
-            moonc check -o ./target/wasm-gc/release/check/char/char.mi -pkg moonbitlang/core/char -i ./target/wasm-gc/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc
-            moonc check ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/check/1/1.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc check ./2/lib.mbt -o ./target/wasm-gc/release/check/2/2.mi -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/check/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc check ./1/lib.mbt ./1/x.wasm-gc.mbt ./1/x_wbtest.wasm-gc.mbt -o ./target/wasm-gc/release/check/1/1.whitebox_test.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm-gc.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm-gc.mbt -o ./target/wasm-gc/release/check/0/0.whitebox_test.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/check/0/0.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["check", "--dry-run", "--target", "js", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc check -o ./target/js/release/check/prelude/prelude.mi -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target js
-            moonc check -o ./target/js/release/check/coverage/coverage.mi -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target js
-            moonc check -o ./target/js/release/check/iter/iter.mi -pkg moonbitlang/core/iter -i ./target/js/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target js
-            moonc check -o ./target/js/release/check/char/char.mi -pkg moonbitlang/core/char -i ./target/js/release/check/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target js
-            moonc check ./1/lib.mbt ./1/x.js.mbt -o ./target/js/release/check/1/1.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js
-            moonc check ./2/lib.mbt -o ./target/js/release/check/2/2.mi -pkg moonbitlang/core/2 -i ./target/js/release/check/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target js
-            moonc check ./1/lib.mbt ./1/x.js.mbt -o ./target/js/release/check/1/1.whitebox_test.mi -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.js.mbt ./0/y_wbtest.js.mbt ./0/y_wbtest.mbt -o ./target/js/release/check/0/0.whitebox_test.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target js -whitebox-test
-            moonc check ./0/lib.mbt ./0/y.js.mbt -o ./target/js/release/check/0/0.mi -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target js
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["build", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/build/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package ./2/lib.mbt -o ./target/wasm-gc/release/build/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/build/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/1/1.core ./target/wasm-gc/release/build/2/2.core -main moonbitlang/core/2 -o ./target/wasm-gc/release/build/2/2.wasm -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/1/1.core -main moonbitlang/core/1 -o ./target/wasm-gc/release/build/1/1.wasm -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/build/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/0/0.core -main moonbitlang/core/0 -o ./target/wasm-gc/release/build/0/0.wasm -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["build", "--dry-run", "--target", "wasm", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/release/build/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm
-            moonc build-package ./2/lib.mbt -o ./target/wasm/release/build/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm/release/build/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm
-            moonc link-core ./target/wasm/release/build/1/1.core ./target/wasm/release/build/2/2.core -main moonbitlang/core/2 -o ./target/wasm/release/build/2/2.wasm -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -target wasm
-            moonc link-core ./target/wasm/release/build/1/1.core -main moonbitlang/core/1 -o ./target/wasm/release/build/1/1.wasm -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -target wasm
-            moonc build-package ./0/lib.mbt ./0/y.wasm.mbt -o ./target/wasm/release/build/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm
-            moonc link-core ./target/wasm/release/build/0/0.core -main moonbitlang/core/0 -o ./target/wasm/release/build/0/0.wasm -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -target wasm
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["build", "--dry-run", "--target", "wasm-gc", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/build/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package ./2/lib.mbt -o ./target/wasm-gc/release/build/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/build/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/1/1.core ./target/wasm-gc/release/build/2/2.core -main moonbitlang/core/2 -o ./target/wasm-gc/release/build/2/2.wasm -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/1/1.core -main moonbitlang/core/1 -o ./target/wasm-gc/release/build/1/1.wasm -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/build/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-            moonc link-core ./target/wasm-gc/release/build/0/0.core -main moonbitlang/core/0 -o ./target/wasm-gc/release/build/0/0.wasm -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["build", "--dry-run", "--target", "js", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt -o ./target/js/release/build/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js
-            moonc build-package ./2/lib.mbt -o ./target/js/release/build/2/2.core -pkg moonbitlang/core/2 -i ./target/js/release/build/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target js
-            moonc link-core ./target/js/release/build/1/1.core ./target/js/release/build/2/2.core -main moonbitlang/core/2 -o ./target/js/release/build/2/2.js -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -target js
-            moonc link-core ./target/js/release/build/1/1.core -main moonbitlang/core/1 -o ./target/js/release/build/1/1.js -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -target js
-            moonc build-package ./0/lib.mbt ./0/y.js.mbt -o ./target/js/release/build/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target js
-            moonc link-core ./target/js/release/build/0/0.core -main moonbitlang/core/0 -o ./target/js/release/build/0/0.js -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -target js
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["test", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/prelude --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/prelude/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -pkg moonbitlang/core/prelude -is-main -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -main moonbitlang/core/prelude -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.wasm -test-mode -pkg-config-path ./prelude/moon.pkg.json -pkg-sources moonbitlang/core/prelude:./prelude -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/iter --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package -o ./target/wasm-gc/debug/test/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0
-            moonc build-package ./target/wasm-gc/debug/test/iter/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/iter/iter.internal_test.core -pkg moonbitlang/core/iter -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/iter/iter.internal_test.core -main moonbitlang/core/iter -o ./target/wasm-gc/debug/test/iter/iter.internal_test.wasm -test-mode -pkg-config-path ./iter/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/iter:./iter -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/coverage --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/coverage/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -pkg moonbitlang/core/coverage -is-main -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -main moonbitlang/core/coverage -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.wasm -test-mode -pkg-config-path ./coverage/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/char --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/char/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/char/char.internal_test.core -pkg moonbitlang/core/char -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/char/char.internal_test.core -main moonbitlang/core/char -o ./target/wasm-gc/debug/test/char/char.internal_test.wasm -test-mode -pkg-config-path ./char/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/char:./char -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/2 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/debug/test/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0
-            moonc build-package ./2/lib.mbt ./target/wasm-gc/debug/test/2/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/2/2.internal_test.core -pkg moonbitlang/core/2 -is-main -i ./target/wasm-gc/debug/test/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.core ./target/wasm-gc/debug/test/2/2.internal_test.core -main moonbitlang/core/2 -o ./target/wasm-gc/debug/test/2/2.internal_test.wasm -test-mode -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind whitebox --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./1/x_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/1/1.whitebox_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.whitebox_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.whitebox_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/1/1.internal_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.internal_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.internal_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind whitebox --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/0/0.whitebox_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/0/0.whitebox_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.whitebox_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/0/0.internal_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/0/0.internal_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.internal_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["test", "--dry-run", "--target", "wasm", "--sort-input"],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/prelude --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./target/wasm/debug/test/prelude/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/prelude/prelude.internal_test.core -pkg moonbitlang/core/prelude -is-main -pkg-sources moonbitlang/core/prelude:./prelude -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/prelude/prelude.internal_test.core -main moonbitlang/core/prelude -o ./target/wasm/debug/test/prelude/prelude.internal_test.wasm -test-mode -pkg-config-path ./prelude/moon.pkg.json -pkg-sources moonbitlang/core/prelude:./prelude -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/iter --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package -o ./target/wasm/debug/test/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm -g -O0
-            moonc build-package ./target/wasm/debug/test/iter/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/iter/iter.internal_test.core -pkg moonbitlang/core/iter -is-main -i ./target/wasm/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/coverage/coverage.core ./target/wasm/debug/test/iter/iter.internal_test.core -main moonbitlang/core/iter -o ./target/wasm/debug/test/iter/iter.internal_test.wasm -test-mode -pkg-config-path ./iter/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/iter:./iter -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/coverage --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./target/wasm/debug/test/coverage/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/coverage/coverage.internal_test.core -pkg moonbitlang/core/coverage -is-main -pkg-sources moonbitlang/core/coverage:./coverage -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/coverage/coverage.internal_test.core -main moonbitlang/core/coverage -o ./target/wasm/debug/test/coverage/coverage.internal_test.wasm -test-mode -pkg-config-path ./coverage/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/char --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./target/wasm/debug/test/char/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/char/char.internal_test.core -pkg moonbitlang/core/char -is-main -i ./target/wasm/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/coverage/coverage.core ./target/wasm/debug/test/char/char.internal_test.core -main moonbitlang/core/char -o ./target/wasm/debug/test/char/char.internal_test.wasm -test-mode -pkg-config-path ./char/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/char:./char -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/2 --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/debug/test/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm -g -O0
-            moonc build-package ./2/lib.mbt ./target/wasm/debug/test/2/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/2/2.internal_test.core -pkg moonbitlang/core/2 -is-main -i ./target/wasm/debug/test/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/1/1.core ./target/wasm/debug/test/2/2.internal_test.core -main moonbitlang/core/2 -o ./target/wasm/debug/test/2/2.internal_test.wasm -test-mode -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm --driver-kind whitebox --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt ./target/wasm/debug/test/1/__generated_driver_for_whitebox_test.mbt -o ./target/wasm/debug/test/1/1.whitebox_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm/debug/test/1/1.whitebox_test.core -main moonbitlang/core/1 -o ./target/wasm/debug/test/1/1.whitebox_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt ./target/wasm/debug/test/1/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/1/1.internal_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/1/1.internal_test.core -main moonbitlang/core/1 -o ./target/wasm/debug/test/1/1.internal_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm --driver-kind whitebox --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm.mbt ./target/wasm/debug/test/0/__generated_driver_for_whitebox_test.mbt -o ./target/wasm/debug/test/0/0.whitebox_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm/debug/test/0/0.whitebox_test.core -main moonbitlang/core/0 -o ./target/wasm/debug/test/0/0.whitebox_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm.mbt ./target/wasm/debug/test/0/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/0/0.internal_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/0/0.internal_test.core -main moonbitlang/core/0 -o ./target/wasm/debug/test/0/0.internal_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["test", "--dry-run", "--target", "wasm-gc", "--sort-input"],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/prelude --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/prelude/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -pkg moonbitlang/core/prelude -is-main -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -main moonbitlang/core/prelude -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.wasm -test-mode -pkg-config-path ./prelude/moon.pkg.json -pkg-sources moonbitlang/core/prelude:./prelude -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/iter --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package -o ./target/wasm-gc/debug/test/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0
-            moonc build-package ./target/wasm-gc/debug/test/iter/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/iter/iter.internal_test.core -pkg moonbitlang/core/iter -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/iter/iter.internal_test.core -main moonbitlang/core/iter -o ./target/wasm-gc/debug/test/iter/iter.internal_test.wasm -test-mode -pkg-config-path ./iter/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/iter:./iter -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/coverage --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/coverage/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -pkg moonbitlang/core/coverage -is-main -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -main moonbitlang/core/coverage -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.wasm -test-mode -pkg-config-path ./coverage/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/char --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./target/wasm-gc/debug/test/char/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/char/char.internal_test.core -pkg moonbitlang/core/char -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/char/char.internal_test.core -main moonbitlang/core/char -o ./target/wasm-gc/debug/test/char/char.internal_test.wasm -test-mode -pkg-config-path ./char/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/char:./char -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/2 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/debug/test/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0
-            moonc build-package ./2/lib.mbt ./target/wasm-gc/debug/test/2/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/2/2.internal_test.core -pkg moonbitlang/core/2 -is-main -i ./target/wasm-gc/debug/test/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.core ./target/wasm-gc/debug/test/2/2.internal_test.core -main moonbitlang/core/2 -o ./target/wasm-gc/debug/test/2/2.internal_test.wasm -test-mode -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind whitebox --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./1/x_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/1/1.whitebox_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.whitebox_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.whitebox_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/1/1.internal_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/1/1.internal_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.internal_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind whitebox --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/0/0.whitebox_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/0/0.whitebox_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.whitebox_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/0/0.internal_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/0/0.internal_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.internal_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["test", "--dry-run", "--target", "js", "--sort-input"],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/prelude --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./target/js/debug/test/prelude/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/prelude/prelude.internal_test.core -pkg moonbitlang/core/prelude -is-main -pkg-sources moonbitlang/core/prelude:./prelude -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/prelude/prelude.internal_test.core -main moonbitlang/core/prelude -o ./target/js/debug/test/prelude/prelude.internal_test.js -test-mode -pkg-config-path ./prelude/moon.pkg.json -pkg-sources moonbitlang/core/prelude:./prelude -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/iter --sort-input --target js --driver-kind internal --mode test
-            moonc build-package -o ./target/js/debug/test/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target js -g -O0
-            moonc build-package ./target/js/debug/test/iter/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/iter/iter.internal_test.core -pkg moonbitlang/core/iter -is-main -i ./target/js/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/coverage/coverage.core ./target/js/debug/test/iter/iter.internal_test.core -main moonbitlang/core/iter -o ./target/js/debug/test/iter/iter.internal_test.js -test-mode -pkg-config-path ./iter/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/iter:./iter -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/coverage --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./target/js/debug/test/coverage/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/coverage/coverage.internal_test.core -pkg moonbitlang/core/coverage -is-main -pkg-sources moonbitlang/core/coverage:./coverage -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/coverage/coverage.internal_test.core -main moonbitlang/core/coverage -o ./target/js/debug/test/coverage/coverage.internal_test.js -test-mode -pkg-config-path ./coverage/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/char --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./target/js/debug/test/char/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/char/char.internal_test.core -pkg moonbitlang/core/char -is-main -i ./target/js/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/coverage/coverage.core ./target/js/debug/test/char/char.internal_test.core -main moonbitlang/core/char -o ./target/js/debug/test/char/char.internal_test.js -test-mode -pkg-config-path ./char/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/char:./char -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/2 --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt -o ./target/js/debug/test/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js -g -O0
-            moonc build-package ./2/lib.mbt ./target/js/debug/test/2/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/2/2.internal_test.core -pkg moonbitlang/core/2 -is-main -i ./target/js/debug/test/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/1/1.core ./target/js/debug/test/2/2.internal_test.core -main moonbitlang/core/2 -o ./target/js/debug/test/2/2.internal_test.js -test-mode -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target js --driver-kind whitebox --mode test
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt ./target/js/debug/test/1/__generated_driver_for_whitebox_test.mbt -o ./target/js/debug/test/1/1.whitebox_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target js -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/js/debug/test/1/1.whitebox_test.core -main moonbitlang/core/1 -o ./target/js/debug/test/1/1.whitebox_test.js -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt ./target/js/debug/test/1/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/1/1.internal_test.core -pkg moonbitlang/core/1 -is-main -pkg-sources moonbitlang/core/1:./1 -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/1/1.internal_test.core -main moonbitlang/core/1 -o ./target/js/debug/test/1/1.internal_test.js -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target js --driver-kind whitebox --mode test
-            moonc build-package ./0/lib.mbt ./0/y.js.mbt ./0/y_wbtest.js.mbt ./0/y_wbtest.mbt ./target/js/debug/test/0/__generated_driver_for_whitebox_test.mbt -o ./target/js/debug/test/0/0.whitebox_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target js -g -O0 -whitebox-test -no-mi
-            moonc link-core ./target/js/debug/test/0/0.whitebox_test.core -main moonbitlang/core/0 -o ./target/js/debug/test/0/0.whitebox_test.js -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./0/lib.mbt ./0/y.js.mbt ./target/js/debug/test/0/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/0/0.internal_test.core -pkg moonbitlang/core/0 -is-main -pkg-sources moonbitlang/core/0:./0 -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/0/0.internal_test.core -main moonbitlang/core/0 -o ./target/js/debug/test/0/0.internal_test.js -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["test", "--dry-run", "--enable-coverage", "--sort-input"],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/prelude --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package -o ./target/wasm-gc/debug/test/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0 -enable-coverage -coverage-package-override=@self
-            moonc build-package ./target/wasm-gc/debug/test/prelude/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -pkg moonbitlang/core/prelude -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/prelude/prelude.internal_test.core -main moonbitlang/core/prelude -o ./target/wasm-gc/debug/test/prelude/prelude.internal_test.wasm -test-mode -pkg-config-path ./prelude/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/prelude:./prelude -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/iter --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package ./target/wasm-gc/debug/test/iter/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/iter/iter.internal_test.core -pkg moonbitlang/core/iter -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/iter/iter.internal_test.core -main moonbitlang/core/iter -o ./target/wasm-gc/debug/test/iter/iter.internal_test.wasm -test-mode -pkg-config-path ./iter/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/iter:./iter -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/coverage --sort-input --target wasm-gc --driver-kind internal --enable-coverage --coverage-package-override=@self --mode test
-            moonc build-package ./target/wasm-gc/debug/test/coverage/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -pkg moonbitlang/core/coverage -is-main -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc -g -O0 -enable-coverage -coverage-package-override=@self -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.internal_test.core -main moonbitlang/core/coverage -o ./target/wasm-gc/debug/test/coverage/coverage.internal_test.wasm -test-mode -pkg-config-path ./coverage/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/char --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package ./target/wasm-gc/debug/test/char/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/char/char.internal_test.core -pkg moonbitlang/core/char -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/char/char.internal_test.core -main moonbitlang/core/char -o ./target/wasm-gc/debug/test/char/char.internal_test.wasm -test-mode -pkg-config-path ./char/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/char:./char -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/2 --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/debug/test/1/1.core -pkg moonbitlang/core/1 -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -enable-coverage
-            moonc build-package ./2/lib.mbt ./target/wasm-gc/debug/test/2/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/2/2.internal_test.core -pkg moonbitlang/core/2 -is-main -i ./target/wasm-gc/debug/test/1/1.mi:1 -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/2:./2 -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/1/1.core ./target/wasm-gc/debug/test/2/2.internal_test.core -main moonbitlang/core/2 -o ./target/wasm-gc/debug/test/2/2.internal_test.wasm -test-mode -pkg-config-path ./2/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/1:./1 -pkg-sources moonbitlang/core/2:./2 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind whitebox --enable-coverage --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./1/x_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/1/1.whitebox_test.core -pkg moonbitlang/core/1 -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -enable-coverage -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/1/1.whitebox_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.whitebox_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/1 --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt ./target/wasm-gc/debug/test/1/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/1/1.internal_test.core -pkg moonbitlang/core/1 -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/1:./1 -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/1/1.internal_test.core -main moonbitlang/core/1 -o ./target/wasm-gc/debug/test/1/1.internal_test.wasm -test-mode -pkg-config-path ./1/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/1:./1 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind whitebox --enable-coverage --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./0/y_wbtest.mbt ./0/y_wbtest.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_whitebox_test.mbt -o ./target/wasm-gc/debug/test/0/0.whitebox_test.core -pkg moonbitlang/core/0 -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -enable-coverage -whitebox-test -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/0/0.whitebox_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.whitebox_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package moonbitlang/core/0 --sort-input --target wasm-gc --driver-kind internal --enable-coverage --mode test
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt ./target/wasm-gc/debug/test/0/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/0/0.internal_test.core -pkg moonbitlang/core/0 -is-main -i ./target/wasm-gc/debug/test/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/0:./0 -target wasm-gc -g -O0 -enable-coverage -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/coverage/coverage.core ./target/wasm-gc/debug/test/0/0.internal_test.core -main moonbitlang/core/0 -o ./target/wasm-gc/debug/test/0/0.internal_test.wasm -test-mode -pkg-config-path ./0/moon.pkg.json -pkg-sources moonbitlang/core/coverage:./coverage -pkg-sources moonbitlang/core/0:./0 -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(&dir, ["bundle", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-            moonc build-package ./2/lib.mbt -o ./target/wasm-gc/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc
-            moonc bundle-core ./target/wasm-gc/release/bundle/0/0.core ./target/wasm-gc/release/bundle/1/1.core ./target/wasm-gc/release/bundle/2/2.core ./target/wasm-gc/release/bundle/coverage/coverage.core ./target/wasm-gc/release/bundle/char/char.core ./target/wasm-gc/release/bundle/iter/iter.core ./target/wasm-gc/release/bundle/prelude/prelude.core -o ./target/wasm-gc/release/bundle/core.core
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["bundle", "--dry-run", "--target", "wasm", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm
-            moonc build-package ./0/lib.mbt ./0/y.wasm.mbt -o ./target/wasm/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm
-            moonc build-package ./2/lib.mbt -o ./target/wasm/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/wasm/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/wasm/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm
-            moonc bundle-core ./target/wasm/release/bundle/0/0.core ./target/wasm/release/bundle/1/1.core ./target/wasm/release/bundle/2/2.core ./target/wasm/release/bundle/coverage/coverage.core ./target/wasm/release/bundle/char/char.core ./target/wasm/release/bundle/iter/iter.core ./target/wasm/release/bundle/prelude/prelude.core -o ./target/wasm/release/bundle/core.core
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["bundle", "--dry-run", "--target", "wasm-gc", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-            moonc build-package ./2/lib.mbt -o ./target/wasm-gc/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc
-            moonc bundle-core ./target/wasm-gc/release/bundle/0/0.core ./target/wasm-gc/release/bundle/1/1.core ./target/wasm-gc/release/bundle/2/2.core ./target/wasm-gc/release/bundle/coverage/coverage.core ./target/wasm-gc/release/bundle/char/char.core ./target/wasm-gc/release/bundle/iter/iter.core ./target/wasm-gc/release/bundle/prelude/prelude.core -o ./target/wasm-gc/release/bundle/core.core
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            ["bundle", "--dry-run", "--target", "js", "--sort-input"],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt -o ./target/js/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js
-            moonc build-package -o ./target/js/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target js
-            moonc build-package ./0/lib.mbt ./0/y.js.mbt -o ./target/js/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target js
-            moonc build-package ./2/lib.mbt -o ./target/js/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/js/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target js
-            moonc build-package -o ./target/js/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/js/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target js
-            moonc build-package -o ./target/js/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/js/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target js
-            moonc build-package -o ./target/js/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target js
-            moonc bundle-core ./target/js/release/bundle/0/0.core ./target/js/release/bundle/1/1.core ./target/js/release/bundle/2/2.core ./target/js/release/bundle/coverage/coverage.core ./target/js/release/bundle/char/char.core ./target/js/release/bundle/iter/iter.core ./target/js/release/bundle/prelude/prelude.core -o ./target/js/release/bundle/core.core
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "bundle",
-                "--target",
-                "all",
-                "--dry-run",
-                "--sort-input",
-                "--serial",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./1/lib.mbt ./1/x.wasm.mbt -o ./target/wasm/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm
-            moonc build-package ./0/lib.mbt ./0/y.wasm.mbt -o ./target/wasm/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm
-            moonc build-package ./2/lib.mbt -o ./target/wasm/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/wasm/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/wasm/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm
-            moonc build-package -o ./target/wasm/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm
-            moonc bundle-core ./target/wasm/release/bundle/0/0.core ./target/wasm/release/bundle/1/1.core ./target/wasm/release/bundle/2/2.core ./target/wasm/release/bundle/coverage/coverage.core ./target/wasm/release/bundle/char/char.core ./target/wasm/release/bundle/iter/iter.core ./target/wasm/release/bundle/prelude/prelude.core -o ./target/wasm/release/bundle/core.core
-            moonc build-package ./1/lib.mbt ./1/x.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target wasm-gc
-            moonc build-package ./0/lib.mbt ./0/y.wasm-gc.mbt -o ./target/wasm-gc/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target wasm-gc
-            moonc build-package ./2/lib.mbt -o ./target/wasm-gc/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/wasm-gc/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/wasm-gc/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target wasm-gc
-            moonc build-package -o ./target/wasm-gc/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target wasm-gc
-            moonc bundle-core ./target/wasm-gc/release/bundle/0/0.core ./target/wasm-gc/release/bundle/1/1.core ./target/wasm-gc/release/bundle/2/2.core ./target/wasm-gc/release/bundle/coverage/coverage.core ./target/wasm-gc/release/bundle/char/char.core ./target/wasm-gc/release/bundle/iter/iter.core ./target/wasm-gc/release/bundle/prelude/prelude.core -o ./target/wasm-gc/release/bundle/core.core
-            moonc build-package ./1/lib.mbt ./1/x.js.mbt -o ./target/js/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target js
-            moonc build-package -o ./target/js/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target js
-            moonc build-package ./0/lib.mbt ./0/y.js.mbt -o ./target/js/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target js
-            moonc build-package ./2/lib.mbt -o ./target/js/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/js/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target js
-            moonc build-package -o ./target/js/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/js/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target js
-            moonc build-package -o ./target/js/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/js/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target js
-            moonc build-package -o ./target/js/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target js
-            moonc bundle-core ./target/js/release/bundle/0/0.core ./target/js/release/bundle/1/1.core ./target/js/release/bundle/2/2.core ./target/js/release/bundle/coverage/coverage.core ./target/js/release/bundle/char/char.core ./target/js/release/bundle/iter/iter.core ./target/js/release/bundle/prelude/prelude.core -o ./target/js/release/bundle/core.core
-            moonc build-package ./1/lib.mbt -o ./target/native/release/bundle/1/1.core -pkg moonbitlang/core/1 -pkg-sources moonbitlang/core/1:./1 -target native
-            moonc build-package -o ./target/native/release/bundle/coverage/coverage.core -pkg moonbitlang/core/coverage -pkg-sources moonbitlang/core/coverage:./coverage -target native
-            moonc build-package ./0/lib.mbt -o ./target/native/release/bundle/0/0.core -pkg moonbitlang/core/0 -pkg-sources moonbitlang/core/0:./0 -target native
-            moonc build-package ./2/lib.mbt -o ./target/native/release/bundle/2/2.core -pkg moonbitlang/core/2 -i ./target/native/release/bundle/1/1.mi:1 -pkg-sources moonbitlang/core/2:./2 -target native
-            moonc build-package -o ./target/native/release/bundle/char/char.core -pkg moonbitlang/core/char -i ./target/native/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/char:./char -target native
-            moonc build-package -o ./target/native/release/bundle/iter/iter.core -pkg moonbitlang/core/iter -i ./target/native/release/bundle/coverage/coverage.mi:coverage -pkg-sources moonbitlang/core/iter:./iter -target native
-            moonc build-package -o ./target/native/release/bundle/prelude/prelude.core -pkg moonbitlang/core/prelude -pkg-sources moonbitlang/core/prelude:./prelude -target native
-            moonc bundle-core ./target/native/release/bundle/0/0.core ./target/native/release/bundle/1/1.core ./target/native/release/bundle/2/2.core ./target/native/release/bundle/coverage/coverage.core ./target/native/release/bundle/char/char.core ./target/native/release/bundle/iter/iter.core ./target/native/release/bundle/prelude/prelude.core -o ./target/native/release/bundle/core.core
-        "#]],
-    );
-}
-
-#[test]
-#[ignore = "not implemented"]
-fn test_backend_flag() {
-    let dir = TestDir::new("backend-flag.in");
-
-    check(
-        get_stdout(&dir, ["check", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc check ./lib/hello.mbt -o ./target/js/release/check/lib/lib.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources username/hello/lib:./lib
-            moonc check ./main/main.mbt -o ./target/js/release/check/main/main.mi -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -i ./target/js/release/check/lib/lib.mi:lib -pkg-sources username/hello/main:./main
-            moonc check ./lib/hello.mbt ./lib/hello_test.mbt -o ./target/js/release/check/lib/lib.underscore_test.mi -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources username/hello/lib:./lib
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["build", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/js/release/build/lib/lib.core -pkg username/hello/lib -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources username/hello/lib:./lib
-            moonc build-package ./main/main.mbt -o ./target/js/release/build/main/main.core -pkg username/hello/main -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -i ./target/js/release/build/lib/lib.mi:lib -pkg-sources username/hello/main:./main
-            moonc link-core $MOON_HOME/lib/core/target/js/release/bundle/core.core ./target/js/release/build/lib/lib.core ./target/js/release/build/main/main.core -main username/hello/main -o ./target/js/release/build/main/main.js -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/main:./main -target js
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["test", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target/js/debug/test --mode test
-            moonc build-package ./lib/hello.mbt ./lib/hello_test.mbt ./target/js/debug/test/lib/__generated_driver_for_underscore_test.mbt -o ./target/js/debug/test/lib/lib.underscore_test.core -pkg username/hello/lib -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources username/hello/lib:./lib -g -ryu
-            moonc link-core $MOON_HOME/lib/core/target/js/release/bundle/core.core ./target/js/debug/test/lib/lib.underscore_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.underscore_test.js -test-mode -pkg-sources username/hello/lib:./lib -target js -ryu
-            moonc build-package ./lib/hello.mbt ./target/js/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources username/hello/lib:./lib -g -ryu
-            moonc link-core $MOON_HOME/lib/core/target/js/release/bundle/core.core ./target/js/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.internal_test.js -test-mode -pkg-sources username/hello/lib:./lib -target js -ryu
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["bundle", "--dry-run", "--sort-input"]),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/js/release/bundle/lib/lib.core -pkg username/hello/lib -pkg-sources username/hello/lib:./lib
-            moonc build-package ./main/main.mbt -o ./target/js/release/bundle/main/main.core -pkg username/hello/main -is-main -i ./target/js/release/bundle/lib/lib.mi:lib -pkg-sources username/hello/main:./main
-            moonc bundle-core ./target/js/release/bundle/lib/lib.core ./target/js/release/bundle/main/main.core -o ./target/js/release/bundle/hello.core
-        "#]],
-    );
-}
-
-#[test]
-fn test_source_map() {
-    let dir = TestDir::new("hello.in");
-
-    // no -source-map in wasm backend
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--target",
-                "wasm",
-                "--debug",
-                "--dry-run",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./main/main.mbt -o ./target/wasm/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm/release/bundle -pkg-sources hello/main:./main -target wasm -g -O0
-            moonc link-core $MOON_HOME/lib/core/target/wasm/release/bundle/core.core ./target/wasm/debug/build/main/main.core -main hello/main -o ./target/wasm/debug/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm -g -O0
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--target",
-                "wasm-gc",
-                "--debug",
-                "--dry-run",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./main/main.mbt -o ./target/wasm-gc/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/wasm-gc/release/bundle -pkg-sources hello/main:./main -target wasm-gc -g -O0 -source-map
-            moonc link-core $MOON_HOME/lib/core/target/wasm-gc/release/bundle/core.core ./target/wasm-gc/debug/build/main/main.core -main hello/main -o ./target/wasm-gc/debug/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target wasm-gc -g -O0 -source-map
-        "#]],
-    );
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--target",
-                "js",
-                "--debug",
-                "--dry-run",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./main/main.mbt -o ./target/js/debug/build/main/main.core -pkg hello/main -is-main -std-path $MOON_HOME/lib/core/target/js/release/bundle -pkg-sources hello/main:./main -target js -g -O0 -source-map
-            moonc link-core $MOON_HOME/lib/core/target/js/release/bundle/core.core ./target/js/debug/build/main/main.core -main hello/main -o ./target/js/debug/build/main/main.js -pkg-config-path ./main/moon.pkg.json -pkg-sources hello/main:./main -pkg-sources moonbitlang/core:$MOON_HOME/lib/core -target js -g -O0 -source-map
-        "#]],
-    );
-}
-
-#[test]
-fn test_find_ancestor_with_mod() {
-    let dir = TestDir::new("hello.in");
-
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            Hello, world!
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["run", "main"]),
-        expect![[r#"
-            Hello, world!
-        "#]],
-    );
-}
-
-#[test]
-fn test_js_format() {
-    let dir = TestDir::new("js_format.in");
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--target",
-                "js",
-                "--sort-input",
-                "--dry-run",
-                "--nostd",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./lib3/hello.mbt -o ./target/js/release/build/lib3/lib3.core -pkg username/hello/lib3 -pkg-sources username/hello/lib3:./lib3 -target js
-            moonc link-core ./target/js/release/build/lib3/lib3.core -main username/hello/lib3 -o ./target/js/release/build/lib3/lib3.js -pkg-config-path ./lib3/moon.pkg.json -pkg-sources username/hello/lib3:./lib3 -target js -exported_functions=hello -js-format iife
-            moonc build-package ./lib2/hello.mbt -o ./target/js/release/build/lib2/lib2.core -pkg username/hello/lib2 -pkg-sources username/hello/lib2:./lib2 -target js
-            moonc link-core ./target/js/release/build/lib2/lib2.core -main username/hello/lib2 -o ./target/js/release/build/lib2/lib2.js -pkg-config-path ./lib2/moon.pkg.json -pkg-sources username/hello/lib2:./lib2 -target js -exported_functions=hello -js-format cjs
-            moonc build-package ./lib1/hello.mbt -o ./target/js/release/build/lib1/lib1.core -pkg username/hello/lib1 -pkg-sources username/hello/lib1:./lib1 -target js
-            moonc link-core ./target/js/release/build/lib1/lib1.core -main username/hello/lib1 -o ./target/js/release/build/lib1/lib1.js -pkg-config-path ./lib1/moon.pkg.json -pkg-sources username/hello/lib1:./lib1 -target js -exported_functions=hello -js-format esm
-            moonc build-package ./lib0/hello.mbt -o ./target/js/release/build/lib0/lib0.core -pkg username/hello/lib0 -pkg-sources username/hello/lib0:./lib0 -target js
-            moonc link-core ./target/js/release/build/lib0/lib0.core -main username/hello/lib0 -o ./target/js/release/build/lib0/lib0.js -pkg-config-path ./lib0/moon.pkg.json -pkg-sources username/hello/lib0:./lib0 -target js -exported_functions=hello -js-format esm
-        "#]],
-    );
-    let _ = get_stdout(&dir, ["build", "--target", "js", "--nostd"]);
-    let t = dir.join("target").join("js").join("release").join("build");
-    check(
-        std::fs::read_to_string(t.join("lib0").join("lib0.js"))
-            .unwrap()
-            .replace_crlf_to_lf(),
-        expect![[r#"
-            function username$hello$lib0$$hello() {
-              return "Hello, world!";
-            }
-            export { username$hello$lib0$$hello as hello }
-        "#]],
-    );
-    check(
-        std::fs::read_to_string(t.join("lib1").join("lib1.js"))
-            .unwrap()
-            .replace_crlf_to_lf(),
-        expect![[r#"
-            function username$hello$lib1$$hello() {
-              return "Hello, world!";
-            }
-            export { username$hello$lib1$$hello as hello }
-        "#]],
-    );
-    check(
-        std::fs::read_to_string(t.join("lib2").join("lib2.js"))
-            .unwrap()
-            .replace_crlf_to_lf(),
-        expect![[r#"
-            function username$hello$lib2$$hello() {
-              return "Hello, world!";
-            }
-            exports.hello = username$hello$lib2$$hello;
-        "#]],
-    );
-    check(
-        std::fs::read_to_string(t.join("lib3").join("lib3.js"))
-            .unwrap()
-            .replace_crlf_to_lf(),
-        expect![[r#"
-            (() => {
-              function username$hello$lib3$$hello() {
-                return "Hello, world!";
-              }
-              globalThis.hello = username$hello$lib3$$hello;
-            })();
         "#]],
     );
 }
@@ -3288,477 +1993,6 @@ fn test_import_shared_memory() {
             moonc build-package ./lib/hello.mbt -o ./target/wasm-gc/release/build/lib/lib.core -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target wasm-gc
             moonc build-package ./main/main.mbt -o ./target/wasm-gc/release/build/main/main.core -pkg username/hello/main -is-main -i ./target/wasm-gc/release/build/lib/lib.mi:lib -pkg-sources username/hello/main:./main -target wasm-gc
             moonc link-core ./target/wasm-gc/release/build/lib/lib.core ./target/wasm-gc/release/build/main/main.core -main username/hello/main -o ./target/wasm-gc/release/build/main/main.wasm -pkg-config-path ./main/moon.pkg.json -pkg-sources username/hello/lib:./lib -pkg-sources username/hello/main:./main -target wasm-gc -import-memory-module xxx -import-memory-name yyy -memory-limits-min 1 -memory-limits-max 65535 -shared-memory
-        "#]],
-    );
-}
-
-#[test]
-fn test_many_targets() {
-    let dir = TestDir::new("test_many_targets.in");
-    check(
-        get_stdout(&dir, ["test", "--target", "all", "--serial"]),
-        expect![[r#"
-            Total tests: 0, passed: 0, failed: 0. [wasm]
-            Total tests: 0, passed: 0, failed: 0. [wasm-gc]
-            Total tests: 0, passed: 0, failed: 0. [js]
-        "#]],
-    );
-
-    check(
-        get_stdout(&dir, ["test", "--target", "js,wasm", "--serial"]),
-        expect![[r#"
-            Total tests: 0, passed: 0, failed: 0. [wasm]
-            Total tests: 0, passed: 0, failed: 0. [js]
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "check",
-                "--target",
-                "js,wasm",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc check ./link/hello.mbt -o ./target/wasm/release/check/link/link.mi -pkg username/hello/link -pkg-sources username/hello/link:./link -target wasm
-            moonc check ./lib/hello.mbt -o ./target/wasm/release/check/lib/lib.mi -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target wasm
-            moonc check ./link/hello.mbt -o ./target/js/release/check/link/link.mi -pkg username/hello/link -pkg-sources username/hello/link:./link -target js
-            moonc check ./lib/hello.mbt -o ./target/js/release/check/lib/lib.mi -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target js
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "build",
-                "--target",
-                "js,wasm",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./link/hello.mbt -o ./target/wasm/release/build/link/link.core -pkg username/hello/link -pkg-sources username/hello/link:./link -target wasm
-            moonc link-core ./target/wasm/release/build/link/link.core -main username/hello/link -o ./target/wasm/release/build/link/link.wasm -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -target wasm
-            moonc build-package ./link/hello.mbt -o ./target/js/release/build/link/link.core -pkg username/hello/link -pkg-sources username/hello/link:./link -target js
-            moonc link-core ./target/js/release/build/link/link.core -main username/hello/link -o ./target/js/release/build/link/link.js -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -target js
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "bundle",
-                "--target",
-                "js,wasm",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moonc build-package ./lib/hello.mbt -o ./target/wasm/release/bundle/lib/lib.core -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target wasm
-            moonc build-package ./link/hello.mbt -o ./target/wasm/release/bundle/link/link.core -pkg username/hello/link -pkg-sources username/hello/link:./link -target wasm
-            moonc bundle-core ./target/wasm/release/bundle/lib/lib.core ./target/wasm/release/bundle/link/link.core -o ./target/wasm/release/bundle/hello.core
-            moonc build-package ./lib/hello.mbt -o ./target/js/release/bundle/lib/lib.core -pkg username/hello/lib -pkg-sources username/hello/lib:./lib -target js
-            moonc build-package ./link/hello.mbt -o ./target/js/release/bundle/link/link.core -pkg username/hello/link -pkg-sources username/hello/link:./link -target js
-            moonc bundle-core ./target/js/release/bundle/lib/lib.core ./target/js/release/bundle/link/link.core -o ./target/js/release/bundle/hello.core
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "test",
-                "--target",
-                "js,wasm",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/wasm/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/wasm/debug/test/link/link.internal_test.wasm -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/js/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/js/debug/test/link/link.internal_test.js -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/js/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.internal_test.js -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "test",
-                "--target",
-                "js,wasm",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-                "-p",
-                "username/hello/lib",
-                "-f",
-                "hello.mbt",
-                "-i",
-                "0",
-            ],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/js/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.internal_test.js -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "test",
-                "--target",
-                "js,wasm,all",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/wasm/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/wasm/debug/test/link/link.internal_test.wasm -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/wasm-gc/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/wasm-gc/debug/test/link/link.internal_test.wasm -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm-gc/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm-gc/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/js/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/js/debug/test/link/link.internal_test.js -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/js/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.internal_test.js -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-        "#]],
-    );
-
-    check(
-        get_stdout(
-            &dir,
-            [
-                "test",
-                "--target",
-                "all",
-                "--dry-run",
-                "--serial",
-                "--nostd",
-                "--sort-input",
-            ],
-        ),
-        expect![[r#"
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/wasm/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/wasm/debug/test/link/link.internal_test.wasm -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm -g -O0 -no-mi
-            moonc link-core ./target/wasm/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/wasm-gc/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/wasm-gc/debug/test/link/link.internal_test.wasm -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target wasm-gc --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/wasm-gc/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/wasm-gc/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target wasm-gc -g -O0 -no-mi
-            moonc link-core ./target/wasm-gc/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/wasm-gc/debug/test/lib/lib.internal_test.wasm -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -target wasm-gc -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/link --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./link/hello.mbt ./target/js/debug/test/link/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/link/link.internal_test.core -pkg username/hello/link -is-main -pkg-sources username/hello/link:./link -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/link/link.internal_test.core -main username/hello/link -o ./target/js/debug/test/link/link.internal_test.js -test-mode -pkg-config-path ./link/moon.pkg.json -pkg-sources username/hello/link:./link -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-            moon generate-test-driver --source-dir . --target-dir ./target --package username/hello/lib --sort-input --target js --driver-kind internal --mode test
-            moonc build-package ./lib/hello.mbt ./target/js/debug/test/lib/__generated_driver_for_internal_test.mbt -o ./target/js/debug/test/lib/lib.internal_test.core -pkg username/hello/lib -is-main -pkg-sources username/hello/lib:./lib -target js -g -O0 -no-mi
-            moonc link-core ./target/js/debug/test/lib/lib.internal_test.core -main username/hello/lib -o ./target/js/debug/test/lib/lib.internal_test.js -test-mode -pkg-config-path ./lib/moon.pkg.json -pkg-sources username/hello/lib:./lib -exported_functions moonbit_test_driver_internal_execute,moonbit_test_driver_finish -js-format cjs -no-dts -target js -g -O0
-        "#]],
-    );
-}
-
-#[test]
-fn test_many_targets_auto_update_001() {
-    let dir = TestDir::new("test_many_targets_auto_update.in");
-    let _ = get_stdout(&dir, ["test", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.wasm.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm")
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("lib").join("x.wasm-gc.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm-gc", content="wasm-gc")
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("lib").join("x.js.mbt")),
-        expect![[r#"
-            test {
-              inspect!("js")
-            }
-        "#]],
-    );
-
-    check(
-        read(dir.join("lib").join("x.native.mbt")),
-        expect![[r#"
-                test {
-                  inspect!("native")
-                }
-            "#]],
-    );
-}
-
-#[test]
-fn test_many_targets_auto_update_002() {
-    let dir = TestDir::new("test_many_targets_auto_update.in");
-    let _ = get_stdout(&dir, ["test", "--target", "js", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.wasm.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm")
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("lib").join("x.wasm-gc.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm-gc")
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("lib").join("x.js.mbt")),
-        expect![[r#"
-            test {
-              inspect!("js", content="js")
-            }
-        "#]],
-    );
-
-    check(
-        read(dir.join("lib").join("x.native.mbt")),
-        expect![[r#"
-            test {
-              inspect!("native")
-            }
-            "#]],
-    );
-
-    let _ = get_stdout(
-        &dir,
-        ["test", "--target", "native", "-u", "--no-parallelize"],
-    );
-    check(
-        read(dir.join("lib").join("x.native.mbt")),
-        expect![[r#"
-            test {
-              inspect!("native", content="native")
-            }
-        "#]],
-    );
-}
-
-#[test]
-fn test_many_targets_auto_update_003() {
-    let dir = TestDir::new("test_many_targets_auto_update.in");
-    let _ = get_stdout(&dir, ["test", "--target", "wasm", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.wasm.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm", content="wasm")
-            }
-        "#]],
-    );
-    check(
-        read(dir.join("lib").join("x.wasm-gc.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm-gc")
-            }
-        "#]],
-    );
-    let _ = get_stdout(&dir, ["test", "--target", "js", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.js.mbt")),
-        expect![[r#"
-            test {
-              inspect!("js", content="js")
-            }
-        "#]],
-    );
-}
-
-#[test]
-fn test_many_targets_auto_update_004() {
-    let dir = TestDir::new("test_many_targets_auto_update.in");
-    let _ = get_stdout(&dir, ["test", "--target", "wasm", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.wasm.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm", content="wasm")
-            }
-        "#]],
-    );
-    let _ = get_stdout(
-        &dir,
-        ["test", "--target", "wasm-gc", "-u", "--no-parallelize"],
-    );
-    check(
-        read(dir.join("lib").join("x.wasm-gc.mbt")),
-        expect![[r#"
-            test {
-              inspect!("wasm-gc", content="wasm-gc")
-            }
-        "#]],
-    );
-    let _ = get_stdout(&dir, ["test", "--target", "js", "-u", "--no-parallelize"]);
-    check(
-        read(dir.join("lib").join("x.js.mbt")),
-        expect![[r#"
-            test {
-              inspect!("js", content="js")
-            }
-        "#]],
-    );
-}
-
-#[test]
-fn test_many_targets_expect_failed() {
-    let dir = TestDir::new("test_many_targets_expect_failed.in");
-    check(
-        get_err_stdout(
-            &dir,
-            ["test", "--target", "all", "--serial", "--sort-input"],
-        ),
-        expect![[r#"
-            test username/hello/lib/x.wasm.mbt::0 failed
-            expect test failed at $ROOT/lib/x.wasm.mbt:2:3-2:32
-            Diff:
-            ----
-            0wasm
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [wasm]
-            test username/hello/lib/x.wasm-gc.mbt::0 failed
-            expect test failed at $ROOT/lib/x.wasm-gc.mbt:2:3-2:35
-            Diff:
-            ----
-            1wasm-gc
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [wasm-gc]
-            test username/hello/lib/x.js.mbt::0 failed
-            expect test failed at $ROOT/lib/x.js.mbt:2:3-2:30
-            Diff:
-            ----
-            2js
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [js]
-        "#]],
-    );
-    check(
-        get_err_stdout(
-            &dir,
-            ["test", "--target", "js,wasm", "--sort-input", "--serial"],
-        ),
-        expect![[r#"
-            test username/hello/lib/x.wasm.mbt::0 failed
-            expect test failed at $ROOT/lib/x.wasm.mbt:2:3-2:32
-            Diff:
-            ----
-            0wasm
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [wasm]
-            test username/hello/lib/x.js.mbt::0 failed
-            expect test failed at $ROOT/lib/x.js.mbt:2:3-2:30
-            Diff:
-            ----
-            2js
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [js]
-        "#]],
-    );
-
-    check(
-        get_err_stdout(
-            &dir,
-            [
-                "test",
-                "--target",
-                "js,wasm,native",
-                "--sort-input",
-                "--serial",
-            ],
-        ),
-        expect![[r#"
-            test username/hello/lib/x.wasm.mbt::0 failed
-            expect test failed at $ROOT/lib/x.wasm.mbt:2:3-2:32
-            Diff:
-            ----
-            0wasm
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [wasm]
-            test username/hello/lib/x.js.mbt::0 failed
-            expect test failed at $ROOT/lib/x.js.mbt:2:3-2:30
-            Diff:
-            ----
-            2js
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [js]
-            test username/hello/lib/x.native.mbt::0 failed
-            expect test failed at $ROOT/lib/x.native.mbt:2:3-2:34
-            Diff:
-            ----
-            3native
-            ----
-
-            Total tests: 1, passed: 0, failed: 1. [native]
         "#]],
     );
 }
