+++ crates/moon/tests/test_cases/targets/auto_update/lib/x.native.mbt
