+++ crates/moon/tests/test_cases/targets/auto_update/moon.mod.json
