+++ crates/moon/tests/test_cases/targets/expect_failed/lib/moon.pkg.json
