+++ crates/moon/tests/test_cases/targets/expect_failed/lib/x.wasm.mbt
