+++ crates/moon/tests/test_cases/targets/expect_failed/moon.mod.json
