+++ crates/moon/tests/test_cases/targets/many_targets/link/moon.pkg.json
